requirements.txt
django
djangorestframework
pillow
psycopg2-binary
python-decouple
python-slugify
